use mime types to determine icon.. if u lok in icons - you'll get he idea - its just mime types exactly splatted out into a dir struct with .db at the end (falling back to default.db and unknonw/unknown/db in the end if it cant fall backto default.db)
now what i need is to talk to cK and get the file magic/mime type stuff to
beocme smarter even that it is.. so currently it sees a tar.gz file - it
looks at the magic and boom.. it thinks is a gzip file.. thats correct.. but
theres mroe to it.. now it woudl be good if the magic stuff coudl now also
inspect the inside of the gzip (ie use zlib in this case) and start lookign
ro a tar header to see if tis a tar.gz.... now if it si a tar.gz.. try
getting the file list and seeing if there are telltale signs of it being a
theme tarball or such (though this lats step may be going too far)

efsd definitely needs ot cache mime type though. that much i'm certain of :)
its not fast at all actually doing file magic on every file... every time
the directory is "loaded" :)

but excellent work! another pat on the back for ck :)


SVN revision: 4451