Convert to using a config dialog instead of config menu.  Courtesy of devilhorns.
This is the patch he posted to the mailing list two days ago.  Committed
after much testing and some discussion with variosu parties.  In
particular, it got a thourough thrashing while we tracked down the
follower confusion bug, as that required constant config changes using
the new dialog.


SVN revision: 18987