The next step.  B-)
Converts fdo menus to directories full of .order files at startup if
there is no applications/menu/all/.order file.  This is the last big
peice of the puzzle.

Lots of tweaking, optimising, cleaning, polishing, and general making
this stuff live well with the rest of E to go.


SVN revision: 25029