Do charset translations to Upper case form.
On, BSD and some applications in Linux the codeset needs to be in the form
listed in 'charsets' manpage.  I.e. utf8 is not good must be UTF-8.  Since
there seems to way to way to get those lists I just added a list in the
code.  It includes all the charsets I can find which are listed in my locale -a.

Let me know if anyone knows how to do this programatically on both Linux and
BSD.


SVN revision: 26445