From: Viktor Kojouharov <vkojouharov@gmail.com>
This little patch adds 'tabs' to the fileman module window. The tabs are
implemented using the toolbar widget. Currently you can add tabs and
switch between them. The only problem right now is that when adding more
than two tabs, the toolbar/scrollbar area is moved off the canvas for
some reason, until the user resizes the window - raster though it might
be a problem with the toolbar widget itself, but he didn't see the
patch, so it might as well be something on my side.



SVN revision: 40306