Option to enable desklock after screensaver timeout.
This option will make it possible to lock the screen after some
seconds screensaver started.

It is useful in cases you want a short screensaver and uses screenlock
but often need to unlock it quickly (ie: was talking to someone at the
office, screensaver starts and you would like to quickly avoid
desklock, otherwise need to stop talking, unlock and continue).



SVN revision: 40413