fix e_border handling of without_border flag.
Ooops! My last changes to introduce without_border broke things, often
unnoticed but it was in the way when we requested resize of existing
window, done in the patch to reuse the e_dialog while toggling between
basic<->advanced by manio.

Thanks manio for spotting and Sachiel for the time help debugging it.


SVN revision: 41313