e_bg: improvements and start import process (unfinished!)
I ran out of time, so if someone wants to finish the move of
e_int_config_wallpaper_import.c to standard e_bg.c I appreciate the
help. Skeleton is there.

With this, it is possible to remove code from conf_wallpaper and have
that functionality in both wallpaper and wallpaper2.



SVN revision: 44527