new, single list module loader dialog.
This dialog was reworked to have one single list instead of two. This
makes the internals way simpler and also fits small screens.

Extra benefit is that we don't need to reload the list, so it's
flicker-free and you don't loose context. Of course it will save some
cpu cycles not having to do these list fills over and over again.

the ilist had to be changed to include an "end" object, in the default
theme it is at the right. It will use a part named
e/widget/ilist/toggle_end to present the selected state. (The
favorites/heart icon was choosen because it is red and thus easily
visible in both black and white background -- selected/unselected).



SVN revision: 46813