postpone module list construction to idler.
Create the list later so the dialog shows up immediately and is later
populated. The creation of the list is quite expensive, with disk,
efreet and later computation of the final list.

It seems good now, but if required it could be split a bit more,
instead of doing one directory per idle run, do one file.



SVN revision: 46843