Fix Window/Actions/{Move to..., Push in Direction...} to account shelf.
Now push to edges or move to position will account for shelves and not
move things below/ontop of them.

Move to help was also changed to say "+" instead of " " as it is more
clear. The code just checks for '-', so anything else is accounted as
positive anyway.



SVN revision: 49432