Window/Action/Push in direction is now able to do 2d movements.
Allow one to send window to corners. up-left, up-right, down-left and
down-right are the new possible values.

HINT: for those that don't know, the "Push" has the nice behavior to
do the push in the given direction but stops in the first
window/obstacle in that direction, thus avoiding overlapping windows.



SVN revision: 49433