From: Peter van de Werken <pwerken-e@A-Eskwadraat.nl>
Attached is a patch that makes the config dialog of the conf_desklock
module work better on small screens.  I moved the advanced dialog
options into a toolbook on the basic dialog.  (I used the Screen Saver
dialog as an example.)
In order to make the wallpaper stuff fit I removed the preview widget.



SVN revision: 49472