Changing action description to the correct parameters.
I'd actually prefer bottom-left and bottom-right instead of down-left
and down-right, but then we would also have to change up-left to top-left
and so on. This change was the one that introduced minimal annoyance to
people who already use this amazing push-to binding (the few people here
who used evilwm should know it).



SVN revision: 53186