Cleanup icon up handling
It seems that originally the "selected" event wasn't to be fired on dirs
during single click. This makes it impossible for fwin to open a new
window when single clicking a dir, so some ugly hacks has been added.
"The right fix (tm)" is to always emit "selected", even on dirs, and let
the listener decide what to do with the file.

SVN revision: 54229