e17 e_border.c: Fixed popup position issue. Closed ticket #627. I tested it and it works fine. Thanks Jim Duchek!
by Jim Duchek <jim.duchek@…> 

After adding a fourth monitor (above my center one) this bug became
utterly unbearable, as it started putting every popup on the bottom of
screens 1-3 as well as to the right on screens 2-4. Anyway, attached
is a patch -- it's a fairly obvious logic error.


SVN revision: 61352