e: actually let E17 handle direct video rendering path.
NOTE: There is still the following list of issue :
	- stacking, with window that could get between the video
layer and the application layer.
	- border, if I turn them off the window get resized to
1x1 and I don't understand yet why.
	- comp-scale, as we don't understand that this window
are one, it display them separatly.

I think that the only possible answer is to enable this feature
only when composite is tunerd on and make window a smart object
so we could merge the video layer and the application in one
object.



SVN revision: 64257