From: Doyoun Kang <doyoun.kang@samsung.com> Subject: [E-devel] [Patch] e_config config save patch
I found a problem in e_config. There is no code to check file size when the e makes the
configuration backup files. Even if the file size is 0, the e copies it to backup file.
So, I added code to check the size of the file. Please check the attached patch file.



SVN revision: 64684