e: fix fps_show in comp being toggled on startup
comp module has an handler on SIGUSR2 to toggle showing fps.
On startup, enlightenment_init sends a SIGUSR2 to its parent so that it
exists of pause() (in e_main.c:537) and finish main init.
Sending SIGUSR1 instead of SIGUSR2 is more of a workaround than a fix.

SVN revision: 67249