Revert a 13 months old commit by englebass
Messagge was:
module unload will delete gadcons, so we need to cleanup in gadman as
well.


This cause all my gadget to be deleted when I switch from 1 monitor
to 2 monitors (while e is running). This probably occcur also when
switching resolution.

Also I don't like all the gadgets configs to be deleted when the 
gadman module is unloaded, we have a conservative approch with
gadget config, so don't see the reason to delete all my gadget
here.

DaveMDS



SVN revision: 67251