econnman: call econnman (app) from the configure/settings.
there is no need to configure it as it will use the EConnMan desktop
file, whatever is available (user can create a new desktop with such
Name= and it will work, this can be done by a gui in e17 itself).

the module will warn people that do not have the application installed.



SVN revision: 76212