efm: have a sane click-to-rename behavior.
Delaying rename entry by 0.75s makes absolutely no sense. As well
there is no sense in always rename if you click the label and the icon
was unselected (quite common in list views).

The proper behavior is to just rename the icon if it was selected, but
ignore the initial click used to select the icon, even if on
label. This is done by remembering the time it was selected, no need
for timer.



SVN revision: 77002