E: Add start of new RandR dialog.
NB: IT IS NOT COMPLETE, DOES NOT APPLY SETTINGS, AND STILL NEEDS LOTS
OF WORK. Don't bother reporting bugs for it just yet, until it is
fully functional. Just adding it to svn in the hopes (hahahaha) that
others may want to jump in and help speed things along.



SVN revision: 77140