E: Fix window hide
The idea of not deleting the borders of unmapped windows started with
r12777 to fix problems when changing virtual desktops. A lot of stuff
was added later to fix this problem, but this original fix has lived on.

When hunting down a mplayer bug raster commented in r14744:
"we need to re-enable the above. when clients hide windows they often
dont destory - keep around to do somethign else with. if we dont unmanage
then we have all sorts of problems believeing it still exists. bigger
problems than mplayer. :)"

So I guess the right thing to do is to destroy the borders of hidden
windows, fixes #917

SVN revision: 77363