e: add a tool to gather statistic.
NOTE: The goal of this tool is to gather statistic about
our user base so we could know what to improve and optimize
in the futur. We don't gather personnal information only
hardware and software configuration, so we know more about
the real use of Enlightenment.

At the moment this tool is not uploading anything on a website
nor is it ready at all. I am just pushing it early so that
people using other OS than Linux, can contribute to it. If that
tool is ready for the release, we will use the gathered statistic
for all our futur plan. If your OS is not in, we will basically
not see it and so not take it into account. That's why I highly
encourage those using E on other OS to contribute to this tool.



SVN revision: 78247