e: Use LT_INIT instead of AC_PROG_LIBTOOL
AC_PROG_LIBTOOL is deprecated and in favor of LT_INIT:

http://www.gnu.org/software/libtool/manual/html_node/LT_005fINIT.html
    Add support for the --enable-shared, --disable-shared, --enable-static,
    --disable-static, --with-pic, and --without-pic configure flags.1
    AC_PROG_LIBTOOL and AM_PROG_LIBTOOL are deprecated names for older
    versions of this macro; autoupdate will upgrade your configure.ac
    files.

With LT_INIT it's possible to shortcut some calls, too. Particularly
disabling support for languages is not needed (only the ones enabled
through  AC_PROG_C* ar enabled) and AC_DISABLE_STATIC is given as an
option.



SVN revision: 78719