e/modules: workaround missing dep in some automake versions
Automake 1.11 seems not to include a dependency between
wizard_page_150_la_OBJECTS and the generated .c/.h files. I'm not
entirely sure this is the right fix, but letting the build broken for
some people is not nice neither. Try to fix it with this, and check
later if this is indeed necessary.



SVN revision: 78998