e: fix relative/absolute path usage in automake
automake thinks it's good to mix absolute and relative path, based on
where the build dir is. So, $(top_srcdir) is not always relative
regardless the fact that if we wanted it to be absolute we could have
used $(abs_top_srcdir). So... use the absolute dir. Always.



SVN revision: 78999