E menu: Fix menu abbreviation.
Possibly not the fastest way to do it, but I really don't have the time to
do it now. I'm just doing it because it took me less time to fix it than
to handle all of Mike's bugging in PM and mails. ;P
Something is still broken in here, but I don't have time to deal with
it, if there's something broken please let me know or fix it yourself,
should be fairly easy.

SVN revision: 79189