e/connman: just update popup list at each 5 secs
If time is not good enough, can be fine tunned
very easily (just change the define).

I couldn't test this commit, so if tasn
or other could, it would be great.

I believe it won't be a perfect solution
yet, since this way it will take some time
to confirm visually the network states
(association / configuration / connect)
in the popup. We need to handle this
case properly (if I'm correct).



SVN revision: 79842